% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bm2meta}
\alias{bm2meta}
\title{bm2meta: simple metadata about bodymap2 samples}
\format{data.frame}
\source{
manually entered, should be checked
}
\usage{
bm2meta
}
\description{
bm2meta: simple metadata about bodymap2 samples
}
\examples{
data(bm2meta)
head(bm2meta)
}
\keyword{datasets}
