% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatherForLocal.R
\name{gatherForLocal}
\alias{gatherForLocal}
\title{naive procedure for copying 18 bodymap2 samples from big RNA compendium}
\usage{
gatherForLocal(destination)
}
\arguments{
\item{destination}{a character(1) identifying a folder where the compendium data will be written by \code{aws s3 cp}.  Approximately 900MB of data will be retrieved.}
}
\description{
naive procedure for copying 18 bodymap2 samples from big RNA compendium
}
\note{
This assumes you have configured AWS credentials to work with compendium.
}
