% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryMeta.R
\name{humrnaMeta}
\alias{humrnaMeta}
\title{arbitrary slice of metadata in public mongodb}
\usage{
humrnaMeta()
}
\description{
arbitrary slice of metadata in public mongodb
}
\examples{
m1 = humrnaMeta()
m1$count()
m1$find(limit=1)
}
