% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryMeta.R
\name{makeRegex}
\alias{makeRegex}
\title{helper for regex construction for mongolite find}
\usage{
makeRegex(fld = "study_center", patt = "UT", opt = "i")
}
\arguments{
\item{fld}{character(1) field name}

\item{patt}{character(1) regex string}

\item{opt}{character(1) option string, 'i' means case-insensitive, an option seems obligatory}
}
\description{
helper for regex construction for mongolite find
}
\examples{
makeRegex()
}
